requirements.txt
tensorflow==2.4.1
tensorflow_probability==0.12.2
tensorflow_privacy==0.5.2
numpy==1.19.5
scipy==1.5.4
scikit_learn==0.23.2
pandas==1.1.5
pandasql==0.7.3
matplotlib==3.3.3
seaborn==0.11.0
faker==5.0.1
simplejson==3.17.2
pyyaml==5.3.1
rstr==2.2.6
rsa==4.7.2
lark-parser==0.11.3
dataclasses-json==0.5.2
dataclasses==0.6;python_version=='3.6'
